InPrayerApp/client/requirements.txt
async-generator==1.10
attrs==23.1.0
beautifulsoup4==4.12.2
black==23.3.0
bs4==0.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
exceptiongroup==1.1.1
h11==0.14.0
idna==3.4
mypy-extensions==1.0.0
numpy==1.24.3
outcome==1.2.0
packaging==23.1
pandas==2.0.2
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.5.3
pycparser==2.21
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
PyQtWebEngine==5.15.7
PyQtWebEngine-Qt5==5.15.2
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.2
requests==2.31.0
selenium==4.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
tomli==2.0.1
trio==0.22.0
trio-websocket==0.10.3
tzdata==2023.3
urllib3==2.0.3
wsproto==1.2.0
